One byte comparison, and it is the careful one
There were four: Hmac.equal and PacketCipher.equalsConstantTime, which did not
exit early, and Connection.sameBytes and KnownHosts.equalBytes, which did.
Nothing at a call site said which had been reached for.

This was survivable only by accident. Both early-exit versions compared host
key blobs, and a host key is something anyone can ask a server for, so learning
one a byte at a time by timing gains nothing. But sameBytes(a, b) does not look
weaker than equal(a, b), and the next thing that needs comparing might be a tag
or a derived key.

So there is one now, in Bytes, and it is constant-time. The cost is a few
nanoseconds on the comparisons that never needed the care, paid on host key
checks that happen once per connection.

Hmac.equal goes rather than moves: nothing called it. The client proves a key
by producing a tag; it is the bridge, which is Python, that has to compare one.
It was written on the argument that the routine belonged beside the primitive,
which turned out to be a reason to write code rather than a use for it.

The tag comparison is the one place here where constant time is load-bearing,
so it is verified where it matters: the full suite still passes against a live
OpenSSH, including the AEAD path that rejects a forged packet.

Closes #54.
